+++ example/lib/main.dart
@@ -10,74 +10,80 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-//
-//import 'package:dart_extensions/dart_extensions.dart';
-//import 'package:flutter/material.dart';
-//
-//void main() => runApp(MyApp());
-//
-//class MyApp extends StatelessWidget {
-//  // This widget is the root of your application.
-//  @override
-//  Widget build(BuildContext context) {
-//    return MaterialApp(
-//      title: 'Flutter Demo',
-//      theme: ThemeData(primarySwatch: Colors.blue),
-//      home: Padding(
-//        padding: const EdgeInsets.all(8.0),
-//        child: Home(),
-//      ),
-//    );
-//  }
-//}
-//
-//class Routes {
-//  static final home = "/home";
-//}
-//
-//class Home extends StatefulWidget {
-//  @override
-//  _HomeState createState() => _HomeState();
-//}
-//
-//class _HomeState extends State<Home> {
-//
-//  navigation() {
-//    navigateTo(route: MaterialPageRoute(builder: (c) => Login()));
-//    navigateByRouteName(Routes.home, );
-//    navigateBack();
-//  }
-//
-//  @override
-//  Widget build(BuildContext context) {
-//
-//    /// For example purpose only
-//
-//    return Scaffold(
-//      body: Container(
-//        child: Stack(
-//          children: <Widget>[
-//            Container(
-//              height: 100,
-//              width: 100,
-//            )   .withRoundCorners(backgroundColor: Colors.grey)
-//                .withShadow()
-//                .alignAtCenter()
-//                .toCenter()
-//                .paddingOnly(left: 10)
-//                .paddingAll(20)
-//                .onTap(() => print('tap'))
-//                .onLongPress(() => print('long press'))
-//          ],
-//        ),
-//      ),
-//    );
-//  }
-//}
-//
-//class Login extends StatelessWidget {
-//  @override
-//  Widget build(BuildContext context) {
-//    return Container();
-//  }
-//}
+
+import 'package:dart_extensions/dart_extensions.dart';
+import 'package:flutter/material.dart';
+
+void main() => runApp(MyApp());
+
+class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      routes: {"/login": (_) => Login()},
+      title: 'Flutter Demo',
+      theme: ThemeData(primarySwatch: Colors.blue),
+      home: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: Home(),
+      ),
+    );
+  }
+}
+
+class Routes {
+  static final login = "/login";
+}
+
+class Home extends StatefulWidget {
+  @override
+  _HomeState createState() => _HomeState();
+}
+
+class _HomeState extends State<Home> {
+  @override
+  Widget build(BuildContext context) {
+    /// For example purpose only
+
+    return Scaffold(
+      body: Container(
+        child: Stack(
+          children: <Widget>[
+            Container(height: 100, width: 100)
+                .withRoundCorners(backgroundColor: Colors.grey)
+                .withShadow()
+                .alignAtCenter()
+                .toCenter()
+                .paddingOnly(left: 10)
+                .paddingAll(20)
+
+                .onTap(
+                  () async {
+                    final result = await navigateByRouteName(Routes.login);
+                    print(result);
+                  },
+                )  .withTooltip('just a tooltip')
+                .onLongPress(() => print('long press'))
+
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class Login extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.grey,
+      child: Center(
+        child: MaterialButton(
+          child: Text('Go Back'),
+          onPressed: () => navigateBack(context: context, result: "hello"),
+        ),
+      ),
+    );
+  }
+}