+++ lib/dart_extensions.dart
@@ -14,19 +14,17 @@
 export 'package:dart_extensions/src/data_stractures/stack.dart';
 export 'package:dart_extensions/src/date.dart';
 export 'package:dart_extensions/src/files.dart';
-export 'package:dart_extensions/src/http.dart';
-export 'package:dart_extensions/src/iterable.dart';
-export 'package:dart_extensions/src/ranges.dart';
-export 'package:dart_extensions/src/int.dart';
-export 'package:dart_extensions/src/search_algorithms.dart';
-export 'package:dart_extensions/src/strings.dart';
-export 'package:dart_extensions/src/sort_algorithms.dart';
 export 'package:dart_extensions/src/flutter/align.dart';
 export 'package:dart_extensions/src/flutter/center.dart';
+export 'package:dart_extensions/src/flutter/container.dart';
 export 'package:dart_extensions/src/flutter/gesture_detector.dart';
 export 'package:dart_extensions/src/flutter/navigation.dart';
 export 'package:dart_extensions/src/flutter/padding.dart';
-export 'package:dart_extensions/src/flutter/container.dart';
-
-
-
+export 'package:dart_extensions/src/http.dart';
+export 'package:dart_extensions/src/int.dart';
+export 'package:dart_extensions/src/iterable.dart';
+export 'package:dart_extensions/src/ranges.dart';
+export 'package:dart_extensions/src/search_algorithms.dart';
+export 'package:dart_extensions/src/sort_algorithms.dart';
+export 'package:dart_extensions/src/strings.dart';
+export 'package:dart_extensions/src/flutter/widgets.dart';