+++ lib/src/flutter/center.dart
@@ -14,11 +14,9 @@
 import 'package:flutter/material.dart';
 
 extension CenterExtension on Widget {
-
   Center toCenter() {
     return Center(
       child: this,
     );
   }
-
 }