+++ lib/src/flutter/container.dart
@@ -12,7 +12,6 @@
  */
 import 'package:flutter/material.dart';
 
-
 extension ContainerExtensions on Container {
   /// Add round corners to a Container
   ///  if the Container already has a color use [backgroundColor] instead and remove the