+++ lib/src/flutter/widgets.dart
@@ -11,24 +11,26 @@
  * limitations under the License.
  */
 
-import 'package:dart_extensions/src/flutter/transforms/click_translate.dart';
 import 'package:flutter/material.dart';
 
-extension GestureDetectorExtensions on Widget {
-  Widget get onTapAddJumpEffect => TranslateOnClick(child: this);
-
-  Widget onDoubleTap(Function function) => GestureDetector(
-        onDoubleTap: function,
-        child: this,
-      );
-
-  Widget onTap(Function function) => GestureDetector(
-        onTap: function,
-        child: this,
-      );
-
-  Widget onLongPress(Function function) => GestureDetector(
-        onLongPress: function,
+extension WidgetsExtension on Widget {
+  Widget withTooltip(String message,
+          {Decoration decoration,
+          double height,
+          bool preferBelow,
+          EdgeInsetsGeometry padding,
+          TextStyle textStyle,
+          Duration waitDuration,
+          EdgeInsetsGeometry margin}) =>
+      Tooltip(
+        message: message,
+        decoration: decoration,
+        height: height,
+        padding: padding,
+        preferBelow: preferBelow,
+        textStyle: textStyle,
+        waitDuration: waitDuration,
+        margin: margin,
         child: this,
       );
 }